+++ bin/generate-load-all-locales.js
@@ -1,17 +1,12 @@
 import path from 'path'
 import fs from 'fs'
-// import javascript_time_ago from 'javascript-time-ago'
-import javascript_time_ago from '../source/JavascriptTimeAgo'
 
 const locales = []
 
-fs.readdirSync(path.resolve(__dirname, '../locale')).forEach(function(locale)
-{
-	const locale_path = path.resolve(__dirname, '../locale', locale)
-	if (fs.statSync(locale_path).isDirectory())
-	{
-		if (fs.existsSync(path.resolve(__dirname, '../locale', locale, 'index.js')))
-		{
+fs.readdirSync(path.resolve(__dirname, '../locale')).forEach((locale) => {
+	const localePath = path.resolve(__dirname, '../locale', locale)
+	if (fs.statSync(localePath).isDirectory()) {
+		if (fs.existsSync(path.resolve(__dirname, '../locale', locale, 'index.js'))) {
 			locales.push(locale)
 		}
 	}