+++ gradation/index.commonjs.js
@@ -1,0 +1,2 @@
+exports = module.exports = require('../commonjs/gradation')
+exports['default'] = require('../commonjs/gradation')