+++ index.commonjs.js
@@ -1,11 +1,9 @@
 'use strict'
 
 exports = module.exports = require('./commonjs/JavascriptTimeAgo').default
-exports['default'] = exports
+exports['default'] = require('./commonjs/JavascriptTimeAgo').default
 
 var locale = require('./commonjs/locale')
 
 exports.intlDateTimeFormatSupported       = locale.intlDateTimeFormatSupported
 exports.intlDateTimeFormatSupportedLocale = locale.intlDateTimeFormatSupportedLocale
-
-exports.RelativeTimeFormat = require('./commonjs/RelativeTimeFormat').default