+++ prop-types/index.commonjs.js
@@ -1,0 +1,2 @@
+exports = module.exports = require('../commonjs/PropTypes')
+exports['default'] = require('../commonjs/PropTypes')