+++ source/PropTypes.js
@@ -1,32 +1,26 @@
-import
-{
+import {
 	oneOfType,
 	arrayOf,
 	string,
 	number,
 	shape,
 	func
-}
-from 'prop-types'
+} from 'prop-types'
 
-const threshold = oneOfType
-([
+const threshold = oneOfType([
 	number,
 	func
 ])
 
-const gradation = arrayOf(oneOfType(
-[
-	shape
-	({
+const gradation = arrayOf(oneOfType([
+	shape({
 		unit        : string.isRequired,
 		factor      : number,
 		granularity : number,
 		threshold
 		// Specific `threshold_[unit]` properties may also be defined
 	}),
-	shape
-	({
+	shape({
 		format : func.isRequired,
 		threshold
 		// Specific `threshold_[unit]` properties may also be defined
@@ -35,23 +29,15 @@ const gradation = arrayOf(oneOfType(
 
 // Date/time formatting style.
 // E.g. 'twitter', 'fuzzy', or custom (`{ gradation: […], units: […], flavour: 'long', custom: function }`)
-const style = oneOfType
-([
+export const style = oneOfType([
 	string,
-	shape
-	({
+	shape({
 		gradation,
 		units   : arrayOf(string),
-		flavour : oneOfType
-		([
+		flavour : oneOfType([
 			string,
 			arrayOf(string)
 		]),
 		custom : func
 	})
 ])
-
-export default
-{
-	style
-}