+++ source/style/default.js
@@ -3,7 +3,7 @@ import { convenient } from '../gradation'
 export default
 {
 	gradation: convenient,
-	flavour: ['long_convenient', 'long'],
+	flavour: ['long-convenient', 'long'],
 	units:
 	[
 		'now',