+++ test/setup.js
@@ -3,16 +3,13 @@ import intlLocalesSupported from 'intl-locales-supported'
 
 import chai, { expect } from 'chai'
 
-import javascript_time_ago from '../source/JavascriptTimeAgo'
+import JavascriptTimeAgo from '../source/JavascriptTimeAgo'
 
 import * as en from '../locale/en'
 import * as ru from '../locale/ru'
 import * as de from '../locale/de'
 import * as ko from '../locale/ko'
 import * as ccp from '../locale/ccp'
-
-// Just so this function code is covered.
-javascript_time_ago.setDefaultLocale('en')
 
 const locales = ['en', 'ru', 'de', 'ko', 'ccp']
 
@@ -43,11 +40,11 @@ else
 // }
 
 // Load localization data for Node.js
-javascript_time_ago.addLocale(en)
-javascript_time_ago.addLocale(ru)
-javascript_time_ago.addLocale(de)
-javascript_time_ago.addLocale(ko)
-javascript_time_ago.addLocale(ccp)
+JavascriptTimeAgo.addLocale(en)
+JavascriptTimeAgo.addLocale(ru)
+JavascriptTimeAgo.addLocale(de)
+JavascriptTimeAgo.addLocale(ko)
+JavascriptTimeAgo.addLocale(ccp)
 
 chai.should()
 global.expect = expect